// src/modules/copastor/types/copastor-form-data.interface.ts
import { type MemberRole } from '@/shared/enums/member-role.enum';
import { MinistryAssignment } from '@/shared/interfaces/ministry-assignment';

export interface CopastorFormData {
  firstNames: string;
  lastNames: string;
  gender: string;
  originCountry: string;
  birthDate: Date;
  maritalStatus: string;
  numberChildren: string;
  conversionDate?: Date;
  email?: string | undefined;
  phoneNumber?: string | undefined;
  residenceCountry: string;
  residenceDepartment: string;
  residenceProvince: string;
  residenceDistrict: string;
  residenceUrbanSector: string;
  residenceAddress: string;
  referenceAddress: string;
  roles: MemberRole[];
  recordStatus?: string | undefined;
  relationType: string;
  theirPastor?: string | undefined;
  theirChurch?: string | undefined;
  theirMinistries?: MinistryAssignment[] | undefined;
}

export type CopastorFormDataKeys =
  | 'firstNames'
  | 'lastNames'
  | 'gender'
  | 'originCountry'
  | 'birthDate'
  | 'maritalStatus'
  | 'numberChildren'
  | 'conversionDate'
  | 'email'
  | 'phoneNumber'
  | 'residenceCountry'
  | 'residenceDepartment'
  | 'residenceProvince'
  | 'residenceDistrict'
  | 'residenceUrbanSector'
  | 'residenceAddress'
  | 'referenceAddress'
  | 'relationType'
  | 'roles'
  | 'recordStatus'
  | 'theirPastor'
  | 'theirChurch'
  | 'theirMinistries';